Create gsva_analysis
Two functions will be required for this analysis.  The first is a function to remove all of the periods off of the transcript read versions so that you can properly merge dataframes.

The second function will be for the creation of the actual expression set that you will use for GSVA analysis.  This expression set has gene names (HUGO-IDs) for rows and samples for columns with TPM data for values.

There are a couple files that are created via R: A file called "target_to_genes" which was created using the sleuth script. This is your translation dictionary, and it is more expansive than going to the ensemble site and downloading the gene names, which can be seen in the intermediate code.

tpm_counts.txt is generated using the "sleuth_gene_table" function and is described in the 2nd half of the r_gsva_analysis. Essentially you run the r_sleuth_pc script and then combine the outcomes, set a q-value to filter by, combine the various conditions you tested for, and then create both tpm_counts.txt as well as sleuth_genes_rank.txt which gets used in the GSVA analysis.